「「更新画面」を作成して「一覧画面」から「更新画面」に遷移して、情報を編集して更新を行う」の実装

+++ app/htdocs/book.php
@@ -4,9 +4,11 @@ declare(strict_types=1);
 require_once(dirname(__DIR__) . "/library/database_access.php");
 
 if(mb_strtolower($_SERVER['REQUEST_METHOD']) === 'post') {
-    $id = $_POST['id'] ?? '';
-    if (!empty($id)) {
-        DatabaseAccess::deleteBy($id);
+    if (isset($_POST['delete'])) {
+        $id = $_POST['id'] ?? '';
+        if (!empty($id)) {
+            DatabaseAccess::deleteBy($id);
+        }
     }
 }
